Reimplent runner using tokio.

+++ src/run.rs
@@ -1,14 +1,19 @@
+#![deny(unused_must_use)]
+
 use errors::*;
-use slog::Logger;
+use futures::{Future, Stream};
+use futures::stream::MergedItem;
 use slog_scope;
-use std::io::{BufRead, BufReader, Read};
-use std::ops::Deref;
+use std::cell::Cell;
+use std::io::{self, BufReader};
 use std::path::Path;
 use std::process::{Command, ExitStatus, Stdio};
-use std::sync::mpsc::{self, RecvTimeoutError};
-use std::sync::mpsc::{Receiver, Sender, channel};
-use std::thread;
+use std::rc::Rc;
 use std::time::Duration;
+use tokio_core::reactor::Core;
+use tokio_io::io::lines;
+use tokio_process::CommandExt;
+use tokio_timer;
 
 pub fn run(name: &str, args: &[&str], env: &[(&str, &str)]) -> Result<()> {
     run_full(None, name, args, env)?;
@@ -93,22 +98,59 @@ const MAX_TIMEOUT_SECS: u64 = 60 * 10 * 2;
 const HEARTBEAT_TIMEOUT_SECS: u64 = 60 * 2;
 
 fn log_command_(mut cmd: Command, capture: bool) -> Result<ProcessOutput> {
-    let mut child = cmd.stdout(Stdio::piped()).stderr(Stdio::piped()).spawn()?;
+    let mut core = Core::new().unwrap();
+    let timer = tokio_timer::wheel()
+        .max_timeout(Duration::from_secs(MAX_TIMEOUT_SECS * 2))
+        .build();
+    let mut child = cmd.stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .spawn_async(&core.handle())?;
 
-    let stdout = child.stdout.take().expect("");
-    let stderr = child.stderr.take().expect("");
+    let stdout = child.stdout().take().expect("");
+    let stderr = child.stderr().take().expect("");
 
     // Needed for killing after timeout
     let child_id = child.id();
 
-    // Child's stdio needs to produce output to avoid being killed
-    let (heartbeat_tx, heartbeat_rx) = mpsc::channel();
+    let heartbeat_timeout = Duration::from_secs(MAX_TIMEOUT_SECS);
 
-    let rx_out = sink(Box::new(stdout),
-                      log_child_stdout,
-                      capture,
-                      heartbeat_tx.clone());
-    let rx_err = sink(Box::new(stderr), log_child_stderr, capture, heartbeat_tx);
+    let heartbeat_timed_out = Rc::new(Cell::new(false));
+
+    let logger = slog_scope::logger();
+    let stdout = lines(BufReader::new(stdout)).map({
+                                                       let logger = logger.clone();
+                                                       move |line| {
+                                                           slog_info!(logger, "blam! {}", line);
+                                                           line
+                                                       }
+                                                   });
+    let stderr = lines(BufReader::new(stderr)).map({
+                                                       let logger = logger.clone();
+                                                       move |line| {
+                                                           slog_info!(logger, "kablam! {}", line);
+                                                           line
+                                                       }
+                                                   });
+    let output = Stream::merge(stdout, stderr);
+    let output = timer
+        .timeout_stream(output, heartbeat_timeout)
+        .map_err({
+                     let cond = heartbeat_timed_out.clone();
+                     move |e| {
+                         if e.kind() == io::ErrorKind::TimedOut {
+                             kill_process(child_id);
+                             cond.set(true);
+                         }
+                         e
+                     }
+                 });
+    let output = if capture {
+        unmerge(output)
+    } else {
+        Box::new(output
+                     .for_each(|_| Ok(()))
+                     .and_then(|_| Ok((Vec::new(), Vec::new()))))
+    };
 
     #[cfg(unix)]
     fn kill_process(id: u32) {
@@ -129,64 +171,26 @@ fn log_command_(mut cmd: Command, capture: bool) -> Result<ProcessOutput> {
         };
     }
 
-    // Have another thread kill the subprocess after the maximum timeout
-    let (timeout_tx, timeout_rx) = mpsc::channel();
-    let (timeout_cancel_tx, timeout_cancel_rx) = mpsc::channel();
-    thread::spawn(move || {
-        let timeout = Duration::from_secs(MAX_TIMEOUT_SECS);
-        match timeout_cancel_rx.recv_timeout(timeout) {
-            Ok(_) => {
-                // Child process exited
-                return;
-            }
-            Err(RecvTimeoutError::Timeout) => {
-                kill_process(child_id);
-                timeout_tx.send(());
-            }
-            _ => panic!(),
-        }
-    });
+    let timed_out = Cell::new(false);
+    let child = timer
+        .timeout(child, Duration::from_secs(MAX_TIMEOUT_SECS))
+        .map_err(|e| {
+                     if e.kind() == io::ErrorKind::TimedOut {
+                         kill_process(child_id);
+                         timed_out.set(true);
+                     }
+                     e
+                 });
 
-    // Have another thread listening for heartbeats on stdout/stderr
-    let (heartbeat_timeout_tx, heartbeat_timeout_rx) = mpsc::channel();
-    let (heartbeat_cancel_tx, heartbeat_cancel_rx) = mpsc::channel();
-    thread::spawn(move || {
-        loop {
-            let timeout = Duration::from_secs(HEARTBEAT_TIMEOUT_SECS);
-            match heartbeat_cancel_rx.recv_timeout(timeout) {
-                Ok(_) => {
-                    // Child process exited
-                    return;
-                }
-                Err(RecvTimeoutError::Timeout) => {
-                    let heartbeats = heartbeat_rx.try_iter().count();
-                    if heartbeats > 0 {
-                        continue;
-                    } else {
-                        // No heartbeats before timeout
-                        kill_process(child_id);
-                        heartbeat_timeout_tx.send(());
-                    }
-                }
-                _ => panic!(),
-            }
-        }
-    });
 
-    let status = child.wait();
-    timeout_cancel_tx.send(());
-    heartbeat_cancel_tx.send(());
-    let timed_out = timeout_rx.try_recv().is_ok();
-    let heartbeat_timed_out = heartbeat_timeout_rx.try_recv().is_ok();
-    let status = status?;
-    let stdout = rx_out.recv().expect("");
-    let stderr = rx_err.recv().expect("");
+    // TODO: Handle errors from tokio_timer better, in particular TimerError::TooLong
+    let (status, (stdout, stderr)) = core.run(Future::join(child, output))?;
 
-    if heartbeat_timed_out {
+    if heartbeat_timed_out.get() {
         info!("process killed after not generating output for {} s",
               HEARTBEAT_TIMEOUT_SECS);
         bail!(ErrorKind::Timeout);
-    } else if timed_out {
+    } else if timed_out.get() {
         info!("process killed after max time of {} s", MAX_TIMEOUT_SECS);
         bail!(ErrorKind::Timeout);
     }
@@ -198,39 +202,21 @@ fn log_command_(mut cmd: Command, capture: bool) -> Result<ProcessOutput> {
        })
 }
 
-fn log_child_stdout(logger: &Logger, line: &str) {
-    slog_info!(logger, "blam! {}", line);
-}
-
-fn log_child_stderr(logger: &Logger, line: &str) {
-    slog_info!(logger, "kablam! {}", line);
-}
-
-fn sink(reader: Box<Read + Send>,
-        log: fn(&Logger, &str),
-        capture: bool,
-        heartbeat_tx: Sender<()>)
-        -> Receiver<Vec<String>> {
-    let (tx, rx) = channel();
-    let logger = slog_scope::logger();
-    thread::spawn(move || {
-        let mut buf = Vec::new();
-        let reader = BufReader::new(reader);
-        for line_bytes in reader.split(b'\n') {
-            if let Ok(line_bytes) = line_bytes {
-                let line = String::from_utf8_lossy(&line_bytes);
-                log(&logger, line.deref());
-                heartbeat_tx.send(());
-                if capture {
-                    buf.push(line.to_string());
-                }
-            } else {
-                log(&logger, "READING FROM CHILD PROCESS FAILED!");
-            }
-        }
-
-        tx.send(buf).expect("");
-    });
-
-    rx
+#[cfg_attr(feature = "cargo-clippy", allow(type_complexity))]
+fn unmerge<T1, T2, S>(reader: S) -> Box<Future<Item = (Vec<T1>, Vec<T2>), Error = S::Error>>
+    where S: Stream<Item = MergedItem<T1, T2>> + 'static,
+          T1: 'static,
+          T2: 'static
+{
+    Box::new(reader
+                 .map(|i| match i {
+                          MergedItem::First(l) => (Some(l), None),
+                          MergedItem::Second(r) => (None, Some(r)),
+                          MergedItem::Both(l, r) => (Some(l), Some(r)),
+                      })
+                 .fold((Vec::new(), Vec::new()), |mut v, i| {
+        i.0.map(|i| v.0.push(i));
+        i.1.map(|i| v.1.push(i));
+        Ok(v)
+    }))
 }
